requirements.txt
torch
torchvision
open_clip_torch
tqdm
Pillow
numpy
kaggle
